# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaApp API",
        "description": "API backend untuk aplikasi InstaApp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments/{postId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "List semua comments pada suatu post",
                "description": "Mengambil semua comments beserta replies untuk post tertentu.",
                "operationId": "5d6254709a3cb617df8ab14e8d58366e",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "UUID post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar comments"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Membuat comment baru pada post",
                "description": "Menambah comment baru atau reply (jika parent_id diisi) untuk post tertentu.",
                "operationId": "9b4dc92dd9b73b156bff880ffc5ebd8b",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "UUID post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Keren banget!"
                                    },
                                    "parent_id": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment berhasil dibuat"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{postId}/{commentId}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Menghapus comment pada post",
                "description": "Menghapus comment utama atau reply berdasarkan UUID.",
                "operationId": "b6fb91f343a6a6b1df59ba879977022a",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "UUID post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "UUID comment",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment berhasil dihapus"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/likes/{postId}": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Like post",
                "description": "Memberikan like pada sebuah post. Jika sudah pernah like, tidak membuat duplikasi.",
                "operationId": "bebbf6e655fab8c10da05cf439c6ddf1",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "UUID dari post yang ingin di-like",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil melakukan like"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Unlike post",
                "description": "Menghapus like dari sebuah post jika sebelumnya sudah like.",
                "operationId": "773cd7a46543334740b4dcf57bd5b10a",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "UUID dari post yang ingin di-unlike",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil melakukan unlike"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}